+++ aws-auth-cognito/src/main/java/com/amplifyframework/auth/cognito/AuthEnvironment.kt
@@ -17,6 +17,7 @@ package com.amplifyframework.auth.cognito
 
 import android.annotation.SuppressLint
 import android.content.Context
+import androidx.core.content.edit
 import com.amplifyframework.auth.cognito.asf.UserContextDataProvider
 import com.amplifyframework.auth.cognito.helpers.SRPHelper
 import com.amplifyframework.auth.exceptions.InvalidStateException
@@ -79,7 +80,7 @@ internal class AuthEnvironment internal constructor(
         val uniqueIdFromPrefs = pinpointPrefs.getString(PINPOINT_UNIQUE_ID_KEY, null)
         val uniqueId = if (uniqueIdFromPrefs == null) {
             val newUniqueId = UUID.randomUUID().toString()
-            pinpointPrefs.edit().putString(PINPOINT_UNIQUE_ID_KEY, uniqueIdFromPrefs).commit()
+            pinpointPrefs.edit(commit = true) { putString(PINPOINT_UNIQUE_ID_KEY, uniqueIdFromPrefs) }
             newUniqueId
         } else {
             uniqueIdFromPrefs